feat: DELETE /challenges/{id} 챌린지 주최자가 아닌 경우 Forbidden 예외처리 추가 #208

+++ src/main/java/com/habitpay/habitpay/domain/challenge/application/ChallengeDeleteService.java
@@ -2,6 +2,8 @@ package com.habitpay.habitpay.domain.challenge.application;
 
 import com.habitpay.habitpay.domain.challenge.dao.ChallengeRepository;
 import com.habitpay.habitpay.domain.challenge.domain.Challenge;
+import com.habitpay.habitpay.global.error.exception.ErrorCode;
+import com.habitpay.habitpay.global.error.exception.ForbiddenException;
 import com.habitpay.habitpay.global.response.SuccessCode;
 import com.habitpay.habitpay.global.response.SuccessResponse;
 import lombok.RequiredArgsConstructor;
@@ -15,8 +17,13 @@ public class ChallengeDeleteService {
     private final ChallengeSearchService challengeSearchService;
     private final ChallengeRepository challengeRepository;
 
-    public SuccessResponse<Void> delete(Long id) {
-        Challenge challenge = challengeSearchService.getChallengeById(id);
+    public SuccessResponse<Void> delete(Long challengeId, Long memberId) {
+        Challenge challenge = challengeSearchService.getChallengeById(challengeId);
+
+        if (challenge.getHost().getId().equals(memberId) == false) {
+            throw new ForbiddenException(ErrorCode.NOT_ALLOWED_TO_DELETE_CHALLENGE);
+        }
+
         challengeRepository.delete(challenge);
 
         return SuccessResponse.of(SuccessCode.DELETE_CHALLENGE_SUCCESS);
